moved RegionInfo to Core namespace; add pcu_type to FindAvailableRegions, + tests
